+++ src/main/java/mech/mania/engine/game/characters/Player.java
@@ -48,8 +48,8 @@ public class Player extends Character {
     }
 
     @Override
-    public double getMaxHealth() {
-        double maxHealth = super.getMaxHealth();
+    public int getMaxHealth() {
+        int maxHealth = super.getMaxHealth();
 
         // Add wearable effects
         if (hat != null) {
@@ -94,8 +94,8 @@ public class Player extends Character {
     }
 
     @Override
-    public double getPhysicalDamage() {
-        double physicalDamage = super.getPhysicalDamage();
+    public int getPhysicalDamage() {
+        int physicalDamage = super.getPhysicalDamage();
 
         // Add wearable effects
         if (hat != null) {
@@ -117,8 +117,8 @@ public class Player extends Character {
     }
 
     @Override
-    public double getMagicalDamage() {
-        double magicalDamage = super.getMagicalDamage();
+    public int getMagicalDamage() {
+        int magicalDamage = super.getMagicalDamage();
 
         // Add wearable effects
         if (hat != null) {
@@ -140,8 +140,8 @@ public class Player extends Character {
     }
 
     @Override
-    public double getMagicalDefense() {
-        double magicalDefense = super.getMagicalDefense();
+    public int getMagicalDefense() {
+        int magicalDefense = super.getMagicalDefense();
 
         // Add wearable effects
         if (hat != null) {
@@ -163,8 +163,8 @@ public class Player extends Character {
     }
 
     @Override
-    public double getPhysicalDefense() {
-        double physicalDefense = super.getPhysicalDefense();
+    public int getPhysicalDefense() {
+        int physicalDefense = super.getPhysicalDefense();
 
         // Add wearable effects
         if (hat != null) {
